// src/app/blog/[slug]/page.module.scss.d.ts
// DO NOT EDIT THIS FILE
// This file is generated automatically by the typed-scss-modules library
export type Styles = {
  authorInde: string
  buttonsWrap: string
  content: string
  contentWrap: string
  main: string
  mainInde: string
  title: string
  xShareButton: string
}

export type ClassNames = keyof Styles

declare const styles: Styles

export default styles